requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
deprecation==2.1.0
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.6
frozenlist==1.5.0
gotrue==2.11.2
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
multidict==6.1.0
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
playwright==1.49.0
postgrest==0.19.3
propcache==0.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyee==12.0.0
pymongo==4.10.1
PyMuPDF==1.25.2
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.2.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
storage3==0.11.1
StrEnum==0.4.15
supabase==2.12.0
supafunc==0.9.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==13.1
yarl==1.18.3
